requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
altgraph==0.17.2
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
audiotsm==0.1.2
av==10.0.0
backoff==2.2.1
beautifulsoup4==4.11.1
blessed==1.20.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.6
cmaes==0.10.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
contourpy==1.1.0
croniter==1.4.1
cssselect==1.2.0
ctranslate2==3.17.1
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.1
Deprecated==1.2.14
docopt==0.6.2
einops==0.6.1
exceptiongroup==1.1.2
fake-useragent==1.1.1
fastapi==0.100.1
faster-whisper==0.7.1
ffmpeg-python==0.2.0
filelock==3.12.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.5.0
fugashi==1.2.1
future==0.18.3
h11==0.14.0
huggingface-hub==0.15.1
humanfriendly==10.0
HyperPyYAML==1.2.1
idna==3.4
importlib-metadata==6.0.0
inquirer==3.1.3
itsdangerous==2.1.2
jaconv==0.3.4
Jinja2==3.1.2
joblib==1.3.1
julius==0.2.7
kiwisolver==1.4.4
lazy_loader==0.3
librosa==0.10.0.post2
lightning==2.0.6
lightning-cloud==0.5.37
lightning-utilities==0.9.0
llvmlite==0.40.1
loguru==0.7.0
lxml==4.9.2
macholib==1.16
Mako==1.2.4
manga-ocr==0.1.10
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mecab-python3==1.0.6
mokuro==0.1.7
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
natsort==8.3.1
networkx==3.1
nltk==3.8.1
numba==0.57.1
numpy==1.22.3
omegaconf==2.3.0
onnxruntime==1.15.1
openai-whisper @ git+https://github.com/openai/whisper.git@b91c907694f96a3fb9da03d4bbdc83fbcd3a40a4
opencv-python==4.5.5.64
optuna==3.2.0
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
parse==1.19.0
Pillow==9.1.0
pooch==1.6.0
primePy==1.3
protobuf==4.23.4
psutil==5.9.5
pyannote.audio==2.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
PyAudio==0.2.13
pyclipper==1.3.0.post4
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pyee==8.2.2
Pygments==2.15.1
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyJWT==2.8.0
pykakasi==2.2.1
pyparsing==3.0.9
pyperclip==1.8.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.6
pytorch-metric-learning==2.3.0
pytube==15.0.0
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
regex==2023.6.3
requests==2.28.2
requests-html==0.10.0
rich==13.4.2
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
scikit-learn==1.3.0
scipy==1.10.1
semver==3.0.1
sentencepiece==0.1.99
shapely==2.0.1
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.5
speechbrain==0.5.15
SpeechRecognition==3.10.0
SQLAlchemy==2.0.19
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.1
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.2
torchmetrics==1.0.1
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.29.2
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
unidic-lite==1.0.8
urllib3==1.26.14
uvicorn==0.23.1
w3lib==2.1.1
wcwidth==0.2.6
websocket-client==1.6.1
websockets==10.4
whisperx==3.1.1
wrapt==1.15.0
yarl==1.9.2
yattag==1.15.1
zipp==3.11.0
